+++ TestApps/SIL.Windows.Forms.TestApp/TestAppForm.cs
@@ -23,6 +23,7 @@ using SIL.Windows.Forms.SettingProtection;
 using SIL.Windows.Forms.WritingSystems;
 using SIL.WritingSystems;
 using SIL.Media;
+using SIL.Windows.Forms.Extensions;
 
 namespace SIL.Windows.Forms.TestApp
 {